% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kc_baltab.R
\name{kc_baltab}
\alias{kc_baltab}
\title{Create balance table with group means and standard deviations or standard error; conduct t-test and F-test}
\usage{
kc_baltab(
  data,
  balvar,
  grpvar,
  refgrp,
  fevar = NULL,
  vcov = "hc1",
  report_sd = FALSE,
  report_ttest_pval = TRUE
)
}
\arguments{
\item{data}{The data frame to tabulate}

\item{balvar}{The variables to tabulate}

\item{grpvar}{If TRUE then drops NA values; if FALSE then includes NA values}

\item{refgrp}{Reference category for group comparisons}

\item{fevar}{Name of fixed effects or block variables to condition on}

\item{vcov}{Standard error type (e.g. "hc1", "iid", "hetero") or a name of a variable to cluster standard errors on}

\item{report_sd}{If TRUE then reports standard deviation of each variable; if FALSE then reports standard errors}

\item{report_ttest_pval}{If TRUE then reports p-values in difference in means test; if FALSE then reports mean difference}
}
\value{
A data frame with tabulation of counts and proportions.
}
\description{
Compute group means, standard deviations, as well as standard errors and conducts difference-in-means test with respect to reference group
}
