% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kc_cor.R
\name{kc_cor}
\alias{kc_cor}
\title{Custom correlation matrix}
\usage{
kc_cor(data, corvar, df = TRUE)
}
\arguments{
\item{data}{The data frame to use}

\item{corvar}{The variables to correlate}

\item{df}{If TRUE then returns output as data frame; if FALSE returns a matrix}
}
\value{
Either a data frame or matrix of correlations.
}
\description{
Compute correlation matrix and display lower triangular and diagonal only.
}
\examples{
\dontrun{
kc_cor(mtcars, c(mpg, disp, wt)) # returns df correlations of mpg, disp, and wt
kc_cor(mtcars, starts_with("c"), df = FALSE) # returns matrix of correlations of vars beginning in c
}
}
