% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kc_figxls.R
\name{kc_figxls}
\alias{kc_figxls}
\title{Export figures to Excel file}
\usage{
kc_figxls(
  plotlist,
  sname = NULL,
  width = 7,
  height = 5.25,
  units = "in",
  fpath = NULL,
  fname = "~/Desktop/KC_Plots.xlsx",
  overwrite = TRUE
)
}
\arguments{
\item{plotlist}{List of plots to export}

\item{sname}{Vector of names for sheet title}

\item{width}{Width}

\item{height}{Height}

\item{units}{Units (units of width and height as in openxlsx::insertPlot())}

\item{fpath}{File path for output; deprecated - use fname exclusively}

\item{fname}{File name for output}

\item{overwrite}{If TRUE then overwrites file; if FALSE then appends sheets on existing file}
}
\description{
Export a list of plots to individual sheets within an Excel file; depends on openxlsx and purrr.
}
