% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kc_regtex.R
\name{kc_regtex}
\alias{kc_regtex}
\title{Export regression table to LaTeX file}
\usage{
kc_regtex(
  reglist,
  sep = FALSE,
  coef_lbl = NULL,
  gof_stat = NULL,
  fpath = "~/Desktop/",
  fname = "mod_all.tex",
  fname_est = "mod_est.tex",
  fname_det = "mod_det.tex"
)
}
\arguments{
\item{reglist}{A list of regression output using base::lm() or fixest::feols()}

\item{sep}{Whether to output coefficient estimates separately from model diagnostics}

\item{coef_lbl}{Coefficient estimates to report; using modelsummary's coef_map argument}

\item{gof_stat}{Vector of diagnostics to be included}

\item{fpath}{File path for output}

\item{fname}{File name for output (both coefficients and diagnostics)}

\item{fname_est}{File name for output (coefficient estimates only)}

\item{fname_det}{File name for output (model diagnostics only)}
}
\description{
Output regression table in tex format with options to separate model estimates from diagnostics; depends on modelsummary, stringr, and purrr.
}
