% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kc_regxls.R
\name{kc_regxls}
\alias{kc_regxls}
\title{Export regression output to Excel file}
\usage{
kc_regxls(
  reglist,
  sname = NULL,
  coef_lbl = NULL,
  gof_stat = NULL,
  rows_lbl = NULL,
  fpath = "~/Desktop/",
  fname = "Regs.xlsx",
  overwrite = TRUE
)
}
\arguments{
\item{reglist}{A list of regression output using base::lm() or fixest::feols()}

\item{sname}{A string for the sheet title}

\item{coef_lbl}{Coefficient estimates to report; using modelsummary's coef_map argument}

\item{gof_stat}{Vector of diagnostics to be included}

\item{rows_lbl}{Additional rows to be included; using modelsummary's add_rows argument}

\item{fpath}{File path for output}

\item{fname}{File name for output}

\item{overwrite}{If TRUE then overwrites file; if FALSE then appends sheet on existing file}
}
\description{
Export a list of plots to individual sheets within an Excel file; depends on modelsummary, openxlsx, and purrr.
}
