% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kc_st_distance.R
\name{kc_st_distance}
\alias{kc_st_distance}
\title{Pairwise distance of sf objects}
\usage{
kc_st_distance(x, y, idx, idy)
}
\arguments{
\item{x}{an sf object}

\item{y}{an sf object}

\item{idx}{id variable for each row in x}

\item{idy}{id variable for each row in y}
}
\value{
A data frame containing pairwise distances of each row in x and y
}
\description{
Create data frame containing pairwise distances of points in sf objects x and y.
}
