% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kc_sum.R
\name{kc_sum}
\alias{kc_sum}
\title{Compute descriptive statistics for provided vars, optionally by group.}
\usage{
kc_sum(data, sumvar, byvar = NULL)
}
\arguments{
\item{data}{The data frame to summarize.}

\item{sumvar}{The variables to summarize.}

\item{byvar}{The variables to group by. (optional)}
}
\value{
A data frame with descriptive statistics: each var is a row; respective stats are in column.
}
\description{
Compute descriptive statistics for provided vars, optionally by group.
}
\examples{
\dontrun{
kc_sum(mtcars, c(mpg, disp))
kc_sum(mtcars, c(mpg, disp), c(vs, am))
kcf_sum(mtcars, where(is.numeric))
}
}
