% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kcf_dup.R
\name{kcf_dup}
\alias{kcf_dup}
\title{Count and tag duplicate rows based on provided vars.}
\usage{
kcf_dup(data, dupvar)
}
\arguments{
\item{data}{The data frame to examine for duplicates}

\item{dupvar}{The variables to check for duplicates}
}
\value{
A data frame with counts per group, tag of whether duplicated, counter within group.
}
\description{
Count and tag duplicate rows based on provided vars.
}
\examples{
\dontrun{
kcf_dup(mtcars, c(vs))
kcf_dup(mtcars, c(vs, am))
}
}
