% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kcf_tab.R
\name{kcf_tab}
\alias{kcf_tab}
\title{Tabulate frequency and relative frequency}
\usage{
kcf_tab(data, tabvar, dropna = FALSE)
}
\arguments{
\item{data}{The data frame to tabulate.}

\item{tabvar}{The variables to tabulate.}
}
\value{
A data frame with tabulation of counts and proportions.
}
\description{
One-way and two-way tabulations of provided variables Speedboost via tidytable.
}
\examples{
\dontrun{
kcf_tab(mtcars, c(vs)) # one-way tabulate
kcf_tab(mtcars, c(vs, am)) # two-way tabulate
}
}
