% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kc_spatial.R
\name{st_pairwise_dist}
\alias{st_pairwise_dist}
\title{Create data frame containing pairwise distances of points in sf objects x and y.}
\usage{
st_pairwise_dist(x, y, idx, idy)
}
\arguments{
\item{x}{an sf object}

\item{y}{an sf object}

\item{idx}{id variable for each row in x}

\item{idy}{id variable for each row in y}
}
\value{
A data frame containing pairwise distances of each row in x and y
}
\description{
Create data frame containing pairwise distances of points in sf objects x and y.
}
