% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kc_theme.R
\name{theme_kc}
\alias{theme_kc}
\title{Theme for ggplot objects (with my preferred defaults)}
\usage{
theme_kc(text_size = 14, text_font = "Arial Narrow")
}
\arguments{
\item{data}{The data frame to tabulate}

\item{tabvar}{The variables to tabulate}

\item{dropna}{If TRUE then drops NA values; if FALSE then includes NA values}
}
\value{
A ggplot2 object.
}
\description{
Similar to hrbrthemes::theme_ipsum()
}
\examples{
\dontrun{
ggplot(mtcars, aes(y = mpg, x = wt, color = factor(am))) +
  geom_point() +
  geom_smooth() +
  labs(x = "Weight", y = "Miles per Gallon", color = "Automatic",
       caption = "Source: Author's calculations using mtcars dataset") +
  theme_kc() +
  theme(legend.position = "bottom")

ggplot(mtcars, aes(y = mpg, x = wt)) +
  geom_point() +
  geom_smooth() +
  facet_wrap(vars(am), scales = "free_x") +
  labs(x = "Weight", y = "Miles per Gallon",
       caption = "Source: Author's calculations using mtcars dataset") +
  theme_kc() +
  theme(legend.position = "bottom")

}
}
