% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kc_dup.R
\name{kc_dup}
\alias{kc_dup}
\title{Indicators for duplicated rows}
\usage{
kc_dup(data, dupvar)
}
\arguments{
\item{data}{The data frame to examine for duplicates}

\item{dupvar}{The variables to check for duplicates}
}
\value{
A data frame with counts per group, tag of whether duplicated, counter within group.
}
\description{
Generate columns with count of occurrences, tag for duplication, and running count of occurrences
}
\examples{
\dontrun{
kc_dup(mtcars, c(vs))
kc_dup(mtcars, c(vs, am))
}
}
