% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kc_regtex.R
\name{kc_regtex}
\alias{kc_regtex}
\title{Custom regression table latex output}
\usage{
kc_regtex(
  reglist,
  sep = FALSE,
  outest = TRUE,
  outdet = TRUE,
  xvar_raw = NULL,
  xvar_lbl = NULL,
  gof_stat = NULL,
  path = "~/Desktop/",
  fname_est = "mod_est.tex",
  fname_det = "mod_det.tex",
  fname = "mod_all.tex"
)
}
\arguments{
\item{reglist}{A list of regression output using base::lm() or fixest::feols()}

\item{sep}{Whether to output coefficients estimates separately from model diagnostics}

\item{outest}{Whether to output coefficients estimates; default is true}

\item{outdet}{Whether to output model diagnostics; default is true}

\item{xvar_raw}{Vector of raw variables names to keep}

\item{xvar_lbl}{Vector of labels for for raw variables to be kept}

\item{gof_stat}{Vector of diagnostics to be included}

\item{path}{File path for output}

\item{fname_est}{File name for output (coefficient estimates only)}

\item{fname_det}{File name for output (model diagnostics only)}

\item{fname}{File name for output (both coefficients and diagnostics)}
}
\description{
Output regression table in latex format, with options for content of output and whether estimates are separate from diagnostics; depends on modelsummary.
}
