% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kc_regxls.R
\name{kc_regxls}
\alias{kc_regxls}
\title{Export regression output to Excel file}
\usage{
kc_regxls(
  reglist,
  sname = NULL,
  coef_map = NULL,
  gof_map = NULL,
  add_rows = NULL,
  add_notes = NULL,
  cols_lbl = NULL,
  mc_cores = 1,
  num_fmt = "\%.3f",
  fpath = NULL,
  fname = "~/Desktop/kc_mod_all.xlsx",
  overwrite = TRUE
)
}
\arguments{
\item{reglist}{A list of regression output using base::lm() or fixest::feols()}

\item{sname}{A string for the sheet title}

\item{coef_map}{Model estimates to report; using modelsummary's coef_map argument}

\item{gof_map}{Model diagnostics to report; using modelsummary's gof_map argument}

\item{add_rows}{Data frame of rows to be appended; using modelsummary's add_rows argument}

\item{add_notes}{A vector of table notes to append at the bottom of the table}

\item{cols_lbl}{A vector of column titles to appear on the top row of the table}

\item{mc_cores}{Parallel computation of model diagnostics; using modelsummary's mc.cores argument}

\item{num_fmt}{Format of numeric values; using modelsummary's fmt argument}

\item{fpath}{File path for output; deprecated - use fname exclusively}

\item{fname}{File name for output}

\item{overwrite}{If TRUE then overwrites file; if FALSE then appends sheet on existing file}
}
\description{
Export a list of regression output to a table within an Excel sheet; depends on modelsummary, openxlsx, dplyr, and purrr.
}
