% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kc_regxls.R
\name{kc_regxls}
\alias{kc_regxls}
\title{Export regression output to Excel file}
\usage{
kc_regxls(
  reglist,
  sname = NULL,
  coef_lbl = NULL,
  gof_lbl = NULL,
  rows_lbl = NULL,
  cols_lbl = NULL,
  note_lbl = NULL,
  num_fmt = "\%.3f",
  fpath = "~/Desktop/",
  fname = "kc_mod_all.xlsx",
  overwrite = TRUE
)
}
\arguments{
\item{reglist}{A list of regression output using base::lm() or fixest::feols()}

\item{sname}{A string for the sheet title}

\item{coef_lbl}{Model estimates to report; using modelsummary's coef_map argument}

\item{gof_lbl}{Model diagnostics to report; using modelsummary's gof_map argument}

\item{rows_lbl}{Table rows to be appended}

\item{cols_lbl}{A vector of column titles to appear on top row of table}

\item{note_lbl}{A vector of table notes to append at the bottom of the table}

\item{num_fmt}{Format of numeric values; using modelsummary's fmt argument}

\item{fpath}{File path for output}

\item{fname}{File name for output}

\item{overwrite}{If TRUE then overwrites file; if FALSE then appends sheet on existing file}
}
\description{
Export a list of plots to individual sheets within an Excel file; depends on modelsummary, openxlsx, dplyr, and purrr.
}
